issue1154: move the resolve methods in Node
