+++ javaparser-core/src/main/java/com/github/javaparser/ast/type/UnknownType.java
@@ -33,6 +33,7 @@ import com.github.javaparser.metamodel.UnknownTypeMetaModel;
 import com.github.javaparser.metamodel.JavaParserMetaModel;
 import javax.annotation.Generated;
 import com.github.javaparser.TokenRange;
+import com.github.javaparser.resolution.types.ResolvedReferenceType;
 import com.github.javaparser.resolution.types.ResolvedType;
 import com.github.javaparser.resolution.types.ResolvedUnionType;
 
@@ -131,6 +132,6 @@ public final class UnknownType extends Type {
 
     @Override
     public ResolvedType resolve() {
-        throw new UnsupportedOperationException();
+        return getSymbolResolver().resolve(this, ResolvedReferenceType.class);
     }
 }
